openstreetmap/pyproject.toml
[project]
name = "osm-dgraph-importer"
version = "0.1.0"
description = "OpenStreetMap to Dgraph knowledge graph importer using OSMnx"
authors = [
    {name = "OSM Import Project", email = ""}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "osmnx>=1.3.0",
    "geopandas>=0.12.2",
    "rdflib>=6.3.2",
    "shapely>=2.0.1",
    "networkx>=3.1",
    "matplotlib>=3.7.1",
    "python-dotenv>=1.0.0",
    "pydgraph>=23.0.0",
    "requests>=2.25.1",
    "pandas>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[project.scripts]
osm-import = "osm_import:main"
osm-validate = "test_osm_data:main"
osm-query = "sample_queries:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
